fix: update setup config

+++ setup/setup.js
@@ -3,15 +3,19 @@ const chalk = require('chalk');
 const prompts = require('prompts');
 const replace = require('replace');
 const rimraf = require('rimraf');
+const util = require('util');
+const exec = util.promisify(require('child_process').exec);
 
 const { log } = require('./constants');
 const questions = require('./setupPrompts');
 const {
     cancelMessage,
+    cleanUpMessage,
     finalMessage,
     gitDeleteMessage,
     gitNoDeleteMessage,
     intalledMessage,
+    pkgAllSetMessage,
     pkgIntroMesage,
 } = require('./messages');
 
@@ -22,15 +26,18 @@ const writeMessage = (msg) => log(chalkBold(msg));
 writeMessage(intalledMessage);
 
 const onCancel = () => {
-    cancelMessage();
+    writeMessage(cancelMessage);
     return false;
 };
+
+const pkgJsonPathPrefix = process.env.MODE === 'test' ? 'setupCopy/' : '';
+const setupPath = process.env.MODE === 'test' ? './setupCopy' : './setup';
 
 // Update package.json
 const updatePackage = async () => {
     writeMessage(pkgIntroMesage);
 
-    const responses = await prompts(questions);
+    const responses = await prompts(questions, { onCancel });
 
     const values = Object.keys(responses).map((item) => ({
         key: item,
@@ -46,7 +53,7 @@ const updatePackage = async () => {
     // update package.json with the user's values
     values.forEach((res) => {
         replace({
-            paths: ['package.json'],
+            paths: [`${pkgJsonPathPrefix}package.json`],
             recursive: false,
             regex: `("${res.key}"): "(.*?)"`,
             replacement: `$1: "${res.value}"`,
@@ -56,7 +63,7 @@ const updatePackage = async () => {
 
     // reset package.json 'keywords' field to empty state
     replace({
-        paths: ['package.json'],
+        paths: [`${pkgJsonPathPrefix}package.json`],
         recursive: false,
         regex: /"keywords": \[[\s\S]+?\]/,
         replacement: `"keywords": []`,
@@ -65,19 +72,62 @@ const updatePackage = async () => {
 
     // remove setup script from package.json
     replace({
-        paths: ['package.json'],
+        paths: [`${pkgJsonPathPrefix}package.json`],
         recursive: false,
         regex: /\s*"setup":.*,/,
         replacement: '',
         silent: true,
     });
 
-    writeMessage(finalMessage);
-
-    // remove all setup scripts from the 'tools' folder
-    rimraf('./setup', (error) => {
-        if (error) throw new Error(error);
+    replace({
+        paths: [`${pkgJsonPathPrefix}package.json`],
+        recursive: false,
+        regex: /\s*"setup:test":.*,/,
+        replacement: '',
+        silent: true,
     });
+
+    replace({
+        paths: [`${pkgJsonPathPrefix}package.json`],
+        recursive: false,
+        regex: /\s*"setup:test:init":.*,/,
+        replacement: '',
+        silent: true,
+    });
+
+    replace({
+        paths: [`${pkgJsonPathPrefix}package.json`],
+        recursive: false,
+        regex: /\s*"setup:copy":.*,/,
+        replacement: '',
+        silent: true,
+    });
+
+    writeMessage(pkgAllSetMessage);
+
+    // message cleanup
+    writeMessage(cleanUpMessage);
+
+    // Remove setup dependencies
+    async function cleanDeps() {
+        const { stdout } = await exec('yarn remove chalk prompts replace -D');
+        console.log('stdout:', stdout);
+    }
+    cleanDeps().then(() => {
+        writeMessage(finalMessage);
+
+        // remove all setup scripts from the 'tools' folder
+        rimraf(setupPath, (rmError) => {
+            if (rmError) throw new Error(rmError);
+        });
+    });
+
+    // writeMessage(finalMessage);
+
+    // // remove all setup scripts from the 'tools' folder
+    // rimraf(setupPath, (rmError) => {
+    //     if (rmError) throw new Error(rmError);
+    // });
 };
 
 // Initialize prompt
@@ -95,12 +145,16 @@ const updatePackage = async () => {
 
     if (deleteGit.value !== undefined) {
         if (deleteGit.value) {
-            rimraf('.git', (error) => {
-                if (error) throw new Error(error);
-
+            if (process.env.MODE === 'test') {
                 writeMessage(gitDeleteMessage);
                 updatePackage();
-            });
+            } else {
+                rimraf('.git', (error) => {
+                    if (error) throw new Error(error);
+                    writeMessage(gitDeleteMessage);
+                    updatePackage();
+                });
+            }
         } else {
             writeMessage(gitNoDeleteMessage);
             updatePackage();
